fix: check to see if the target is "concat"-ing
Targets specified as `* / foo`, pointing at a sub-router, are intended
to act like concatenation, not nesting. If we're in that mode, preserve
the consumed `/` from the route.

If we're not strictly concatenating, only add the last slash if we've
completely consumed the route (to allow for sub-routers to mount a view
at `/`.)